drivers/s3: hazard a guess at region field from URL
Look, is selecting the second segment of the domain name kind of
arbitrary? Yeah. Should we also build in the ability to customize the
region in from the OS string? Probably. But it's not like any OSses
actually _require_ the string "ignored", so I'd argue this change does
no harm while "magically" making things start working for Wasabi users.
Also anyone who uses the s3+https syntax for first-party S3 for whatever
reason.